refactor apply_hf_quant
