+++ mlx_lm/quant/utils.py
@@ -34,10 +34,10 @@ def load_data(tokenizer, num_samples: int, sequence_length: int) -> mx.array:
 
 def replace_linear_with_quant_linear(model, quant_method = "bitnet", modules_to_not_convert=None):
     quantize_layers = []
-    for name, module in model.named_modules():     
+    for name, module in model.named_modules():
         if modules_to_not_convert is None:
             modules_to_not_convert = []
-        
+
         # Replace nn.Linear layers, but skip 'lm_head'
         if name not in modules_to_not_convert and isinstance(module, nn.Linear):
             old_weight = module.weight
@@ -50,4 +50,23 @@ def replace_linear_with_quant_linear(model, quant_method = "bitnet", modules_to_
             quantize_layers.append((name, new_layer))
     if len(quantize_layers) > 0:
         model.update_modules(tree_unflatten(quantize_layers))
+    return model
+
+
+def apply_hf_quantization(model, config):
+    """
+    Apply HF quantization to a model if it has a quantization config.
+    """
+    if config.get("quantization_config", None) is not None:
+        quantization_config = config["quantization_config"]
+        quant_method = quantization_config.get("quant_method", None)
+        modules_to_not_convert = quantization_config.get("modules_to_not_convert", None)
+
+        if quant_method is not None and quant_method in QUANT_LINEAR_MAPPING.keys():
+            # Replace linear layers with quantized versions
+            model = replace_linear_with_quant_linear(
+                model,
+                quant_method=quant_method,
+                modules_to_not_convert=modules_to_not_convert
+            )
     return model
